+++ examples/example_taggged_signals.rs
@@ -113,7 +113,7 @@ fn main() -> Result<(), Box<dyn std::error::Error>> {
     let factories = Factories::read(std::io::BufReader::new(std::fs::File::open("examples/example.glade")?))?;
 
     gtk::init()?;
-    woab::run_actix_inside_gtk_event_loop()?;
+    woab::run_actix_inside_gtk_event_loop();
 
     woab::block_on(async {
         factories.win_app.instantiate().connect_with(|bld| {